requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
tableauserverclient>=0.25.0
python-dotenv>=1.0.0
defusedxml>=0.7.1
